adicionando altas coisa

+++ src/configuration/erros/erros.go
@@ -7,7 +7,7 @@ import (
 type RestErr struct {
 	Message string   `json:"message"`
 	Err     string   `json:"error"`
-	Code    int64    `json:"code"`
+	Code    int      `json:"code"`
 	Causes  []Causes `json:"causes"`
 }
 
@@ -16,7 +16,7 @@ type Causes struct {
 	Message string `json:"message"`
 }
 
-func NewRestERR(message, err string, code int64, causes []Causes) *RestErr {
+func NewRestERR(message, err string, code int, causes []Causes) *RestErr {
 	return &RestErr{
 		Message: message,
 		Err:     err,
